first public release of CA toolset and docs
These tools have been used internally for a few years. The public
release contains a few small bug fixes, editing of the documentation,
a "desk check" of the example, and the removal of proprietary information
references.

The largest additions are the "example" scripts that capture the
steps in the example (see "tools/README.md").