initca tool, expanded docs, new and renamed tools (#2)
The process of setting up the CA has been cleaned up a bit with a new tool to automate the CA initialisation. The tool uses template/prototype files to generate conf files when initialising the CA. All this is explained in the docs and scripts.

Some tools have been renamed for consistency reasons (e.g. consistent use of "mk" for making things). Tools to support certification revocation and revocation lists have been added, but they are not as mature as the previous tools.

With the tool changes, the README doc and the long example have been re-edited and hopefully improved. The code in the README can be found in the example folder, but the names have been changed (a numeric prefix added) so one can see the sequence.